pyproject.toml
[tool.poetry]
name = "djw"
version = "0.1.0"
description = ""
authors = ["roma <roma.marhatok@yandex.by>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
psycopg2-binary = "^2.9.3"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
flake8-bugbear = "^22.4.25"
flake8-builtins = "^1.5.3"
pep8-naming = "^0.12.1"
flake8-variables-names = "^0.0.5"
autopep8 = "^1.6.0"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
